requirements.txt
Flask
Flask-Cors
Flask-SQLAlchemy
psycopg2-binary
SQLAlchemy
python-dotenv
psycopg2
Flask-Migrate
flask-bcrypt
mysql-connector
mysql-connector-python
mysql-connector-python-rf
pymysql
cryptography